pkg/system: fix compile on darwin (macOS)
