+++ pkg/system/chtimes_linux_test.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package system // import "github.com/docker/docker/pkg/system"
 
 import (
