+++ pkg/system/rm.go
@@ -1,4 +1,4 @@
-// +build !windows
+// +build !darwin,!windows
 
 package system // import "github.com/docker/docker/pkg/system"
